Let each player have its own color

+++ backend/src/main/java/com/example/backend/GameSupportService.java
@@ -53,7 +53,7 @@ public class GameSupportService extends TextWebSocketHandler {
                 gameRulesService.reset();
                 break;
             case "rollDice":
-                var playerNumber = session.getAttributes().get("playerNumber");
+                var playerNumber = (Integer) session.getAttributes().get("playerNumber");
                 if (playerNumber.equals(gameRulesService.getGameStatus().activePlayerNumber())) {
                     gameRulesService.rollDice(playerNumber);
                 }
